requirements.txt
fastapi
uvicorn
pydantic>=2.0,<3.0
gradio
requests
langchain>=0.2.0
langchain-community
langchain-core
langchain-openai
langchain-chroma
langchain-huggingface
langchain-google-genai
langgraph
chromadb>=0.4.0
sentence-transformers
neo4j>=5.0.0
pymupdf>=1.23.0
llama-parse>=0.4.0
python-dotenv
pytest>=8.0.0
pytest-cov
